Authorization
The user now has to be registered and logged in to create new posts and categories. Added they database column to store the user id so that only the user that created the specific post and category can edit or delete it.

+++ app/Http/Controllers/CategoryController.php
@@ -25,27 +25,43 @@ class CategoryController extends Controller
             'name' => 'required',
         ]);
 
+
+        $formFields['user_id'] = auth()->id();
+
         Category::create($formFields);
 
         return redirect('/categories');
     }
 
-    public function edit(Category $category) {
-        return view('categories.edit',['category' => $category]);
+    public function edit(Category $category)
+    {
+        return view('categories.edit', ['category' => $category]);
     }
 
-     public function update(Request $request, Category $category)
+    public function update(Request $request, Category $category)
     {
+        if ($category->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action!');
+        }
+
         $formFields = $request->validate([
             'name' => 'required',
         ]);
+
+        $formFields['user_id'] = auth()->id();
 
         $category->update($formFields);
 
         return back();
     }
 
-    public function destroy(Category $category){
+    public function destroy(Category $category)
+    {
+
+        if ($category->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action!');
+        }
+
         $category->delete();
 
         return redirect('/categories');
